requirements.txt.txt
gspread
oauth2client
python-telegram-bot==13.15
